Se modifican MODEL / SERVICES / CONTROLLER  de backend

+++ backend/woofandbarf/src/main/java/org/proyecto_integrador/woofandbarf/repository/CartRepository.java
@@ -4,7 +4,6 @@ import org.proyecto_integrador.woofandbarf.model.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-// Nota: El segundo parámetro es Integer (el tipo del ID)
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
 }
